[Feature] Всё, но без защиты и подмены
