+++ src/components/profile-form/profile-form.module.css
@@ -1,0 +1,34 @@
+.form {
+  display: flex;
+  flex-flow: column nowrap;
+  justify-content: flex-start;
+  align-items: center;
+  width: 480px;
+  }
+
+.form__fieldset {
+  border: none;
+  width: 100%;
+  align-items: center;
+  box-sizing: border-box;
+  }
+.form__fieldset > div {
+  padding: 12px 0;
+  }
+
+.form__fieldset > div > div  {
+  width: 480px;
+  height: 64px;
+  }
+
+.form__stub {
+  min-height: 24px;
+  }
+
+.profile__buttons {
+  width: 480px;
+  display: flex;
+  flex-flow: row nowrap;
+  justify-content: space-evenly;
+  align-items: center;
+  }
