+++ src/components/profile-orders/profile-orders.jsx
@@ -1,0 +1,5 @@
+import { React } from 'react';
+
+const ProfileOrders = () => (<h1 className='text text_type_main-large'>Страница в разработке!</h1>);
+
+export default ProfileOrders;
