+++ src/pages/profile-page/profile-page.module.css
@@ -4,71 +4,4 @@
   gap: 0 60px;
   overflow: hidden;
   height: 100vh;
-  padding: 40px 340px;
   }
-
-.menu {
-  display: flex;
-  flex-flow: column nowrap;
-  justify-content: flex-start;
-  align-items: flex-start;
-
-  }
-.menu__item {
-  width:  320px;
-  height: 64px;
-  text-decoration: none;
-  color: inherit;
-  vertical-align: center;
-  display: flex;
-  flex-flow: column;
-  justify-content: center;
-  }
-.menu__item_type_button {
-  background-color: transparent;
-  border: none;
-  text-align: left;
-  }
-
-.form {
-  display: flex;
-  flex-flow: column nowrap;
-  justify-content: flex-start;
-  align-items: center;
-  width: 480px;
-  padding-top: 120px;
-  }
-
-.form__heading {
-  margin-block-start: 0;
-  margin-block-end: 0;
-  margin-inline-start: 0;
-  margin-inline-end: 0;
-  }
-
-.form__fieldset {
-  border: none;
-  width: 100%;
-  align-items: center;
-  box-sizing: border-box;
-  }
-.form__fieldset > div {
-  padding: 12px 0;
-  }
-
-.form__fieldset > div > div  {
-  width: 480px;
-  height: 64px;
-  }
-
-.form__stub {
-  min-height: 24px;
-  }
-
-.profile__buttons {
-  width: 480px;
-  display: flex;
-  flex-flow: row nowrap;
-  justify-content: space-evenly;
-  align-items: center;
-}
